Moved Menubar component into navigator, moved MenubarItem into new file, renamed components into MenuBar and MenuBarItem
Auditors: @bsclifton

Test Plan:
- open main menu in windows
- navigate trough items with keys
- open some items
